Moved Type+SwiftSyntax to separate file
