+++ SourceryFramework/Sources/Parsing/SwiftSyntax/AST/TypeName+SwiftSyntax.swift
@@ -19,15 +19,6 @@ extension SyntaxProtocol {
     }
 }
 
-extension Type {
-    convenience init?(_ node: TypeSyntax) {
-        guard let typeIdentifier = node.as(SimpleTypeIdentifierSyntax.self) else { return nil }
-        let name = typeIdentifier.name.text.trimmed
-        let generic = typeIdentifier.genericArgumentClause.map { GenericType(name: typeIdentifier.name.text, node: $0) }
-        self.init(name: name, isGeneric: generic != nil)
-    }
-}
-
 extension TypeName {
     convenience init(_ node: TypeSyntax) {
         /* TODO: redesign what `TypeName` represents, it can represent all those different variants
